BITS-1856: Eledger config BE - Implement GL account's persistence model

+++ src/main/java/bhutan/eledger/application/port/in/config/balanceaccount/CreateBalanceAccountPartTypeUseCase.java
@@ -12,7 +12,7 @@ import java.util.Map;
 @Validated
 public interface CreateBalanceAccountPartTypeUseCase {
 
-    Long create(@Valid CreateBalanceAccountPartTypeUseCase.CreateBalanceAccountPartTypeCommand command);
+    Integer create(@Valid CreateBalanceAccountPartTypeUseCase.CreateBalanceAccountPartTypeCommand command);
 
     @Data
     class CreateBalanceAccountPartTypeCommand {
